requirements/prompt_classifier_requirements.txt
# Privacy Guard LLM - Multi-LLM Prompt Classifier Requirements

# Core dependencies
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# LLM API clients
openai>=1.3.0
anthropic>=0.8.0
google-generativeai>=0.3.0
cohere>=4.37.0

# Data analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Logging and monitoring
loguru>=0.7.0
tqdm>=4.66.0

# Utilities
requests>=2.31.0
jsonschema>=4.17.0
python-dateutil>=2.8.0
pytz>=2023.3

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0